backend/requirements.txt
Django==5.0.1
whitenoise==6.11.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
